combine all enum into 1 file

+++ lib/features/players/widgets/squad_player_tile.dart
@@ -1,8 +1,8 @@
 import 'package:flutter/material.dart';
 import 'package:tracket/common/widgets/image_circle_avatar.dart';
 import 'package:tracket/features/players/screens/player_profile_screen.dart';
-import 'package:tracket/features/teams/models/team_role.dart';
 import 'package:tracket/utils/constants/colors.dart';
+import 'package:tracket/utils/constants/enums.dart';
 import 'package:tracket/utils/helpers/helping_function.dart';
 
 class SquadPlayerTile extends StatelessWidget {
@@ -36,9 +36,7 @@ class SquadPlayerTile extends StatelessWidget {
     return Card(
       elevation: 0,
       margin: const EdgeInsets.symmetric(horizontal: 0, vertical: 2),
-      color: isDark
-          ? DarkThemeColors.secondaryBackground
-          : LightThemeColors.secondaryBackground,
+      color: isDark ? DarkThemeColors.secondaryBackground : LightThemeColors.secondaryBackground,
       shape: RoundedRectangleBorder(
         borderRadius: BorderRadius.circular(12),
         side: BorderSide(
@@ -47,8 +45,7 @@ class SquadPlayerTile extends StatelessWidget {
       ),
       child: InkWell(
         borderRadius: BorderRadius.circular(12),
-        onTap: () => THelperFunction.pushScreen(
-            context, PlayerProfileScreen(playerId: playerId)),
+        onTap: () => THelperFunction.pushScreen(context, PlayerProfileScreen(playerId: playerId)),
         child: Padding(
           padding: const EdgeInsets.all(12),
           child: Row(
@@ -63,8 +60,7 @@ class SquadPlayerTile extends StatelessWidget {
                   isWicketKeeper: isWicketKeeper,
                 ),
               ),
-              if (isEdit && teamRole != TeamRole.owner)
-                _DeleteButton(onDelete: onDelete),
+              if (isEdit && teamRole != TeamRole.owner) _DeleteButton(onDelete: onDelete),
             ],
           ),
         ),
@@ -149,17 +145,13 @@ class _PlayerInfo extends StatelessWidget {
             if (isCaptain)
               _RoleLabel(
                 label: 'Captain',
-                color: isDark
-                    ? DarkThemeColors.batsmanColor
-                    : LightThemeColors.batsmanColor,
+                color: isDark ? DarkThemeColors.batsmanColor : LightThemeColors.batsmanColor,
                 icon: Icons.star_rounded,
               ),
             if (isWicketKeeper)
               _RoleLabel(
                 label: 'Wicketkeeper',
-                color: isDark
-                    ? DarkThemeColors.wicketKeeperColor
-                    : LightThemeColors.wicketKeeperColor,
+                color: isDark ? DarkThemeColors.wicketKeeperColor : LightThemeColors.wicketKeeperColor,
                 icon: Icons.sports_cricket_rounded,
               ),
           ],
